+++ dflib/src/main/java/org/dflib/exp/bool/BoolColumn.java
@@ -1,11 +1,10 @@
 package org.dflib.exp.bool;
 
 import org.dflib.BooleanSeries;
-import org.dflib.BoolValueMapper;
+import org.dflib.Condition;
 import org.dflib.DataFrame;
 import org.dflib.Series;
 import org.dflib.exp.Column;
-import org.dflib.Condition;
 
 
 public class BoolColumn extends Column<Boolean> implements Condition {
@@ -26,16 +25,12 @@ public class BoolColumn extends Column<Boolean> implements Condition {
     @Override
     public BooleanSeries eval(DataFrame df) {
         Series<Boolean> c = super.eval(df);
-        return c instanceof BooleanSeries
-                ? (BooleanSeries) c
-                : c.mapAsBool(BoolValueMapper.of());
+        return c.compactBool();
     }
 
     @Override
     public BooleanSeries eval(Series<?> s) {
         Series<Boolean> c = super.eval(s);
-        return c instanceof BooleanSeries
-                ? (BooleanSeries) c
-                : c.mapAsBool(BoolValueMapper.of());
+        return c.compactBool();
     }
 }