+++ dflib/src/main/java/org/dflib/slice/DeferredColumnSet.java
@@ -6,13 +6,12 @@ import org.dflib.ColumnDataFrame;
 import org.dflib.ColumnSet;
 import org.dflib.Condition;
 import org.dflib.DataFrame;
-import org.dflib.DoubleSeries;
 import org.dflib.DoubleValueMapper;
 import org.dflib.Exp;
+import org.dflib.FloatValueMapper;
 import org.dflib.Index;
 import org.dflib.IntSeries;
 import org.dflib.IntValueMapper;
-import org.dflib.LongSeries;
 import org.dflib.LongValueMapper;
 import org.dflib.RowColumnSet;
 import org.dflib.RowMapper;
@@ -207,7 +206,7 @@ public class DeferredColumnSet implements ColumnSet {
 
         for (int i = 0; i < w; i++) {
             Series s = source.getColumn(i);
-            columns[i] = s instanceof BooleanSeries ? s.castAsBool() : s.mapAsBool(BoolValueMapper.of());
+            columns[i] = s.compactBool();
         }
 
         return new ColumnDataFrame(null, source.getColumnsIndex(), columns);
@@ -220,7 +219,7 @@ public class DeferredColumnSet implements ColumnSet {
 
         for (int i = 0; i < w; i++) {
             Series s = source.getColumn(i);
-            columns[i] = s.mapAsBool(mapper);
+            columns[i] = s.compactBool(mapper);
         }
 
         return new ColumnDataFrame(null, source.getColumnsIndex(), columns);
@@ -231,11 +230,9 @@ public class DeferredColumnSet implements ColumnSet {
         int w = source.width();
         Series<?>[] columns = new Series[w];
 
-        IntValueMapper<?> mapper = IntValueMapper.of(forNull);
-
         for (int i = 0; i < w; i++) {
             Series s = source.getColumn(i);
-            columns[i] = s instanceof IntSeries ? s.castAsInt() : s.mapAsInt(mapper);
+            columns[i] = s.compactInt(forNull);
         }
 
         return new ColumnDataFrame(null, source.getColumnsIndex(), columns);
@@ -248,7 +245,7 @@ public class DeferredColumnSet implements ColumnSet {
 
         for (int i = 0; i < w; i++) {
             Series s = source.getColumn(i);
-            columns[i] = s.mapAsInt(mapper);
+            columns[i] = s.compactInt(mapper);
         }
 
         return new ColumnDataFrame(null, source.getColumnsIndex(), columns);
@@ -259,11 +256,9 @@ public class DeferredColumnSet implements ColumnSet {
         int w = source.width();
         Series<?>[] columns = new Series[w];
 
-        LongValueMapper<?> mapper = LongValueMapper.of(forNull);
-
         for (int i = 0; i < w; i++) {
             Series s = source.getColumn(i);
-            columns[i] = s instanceof LongSeries ? s.castAsLong() : s.mapAsLong(mapper);
+            columns[i] = s.compactLong(forNull);
         }
 
         return new ColumnDataFrame(null, source.getColumnsIndex(), columns);
@@ -276,7 +271,33 @@ public class DeferredColumnSet implements ColumnSet {
 
         for (int i = 0; i < w; i++) {
             Series s = source.getColumn(i);
-            columns[i] = s.mapAsLong(mapper);
+            columns[i] = s.compactLong(mapper);
+        }
+
+        return new ColumnDataFrame(null, source.getColumnsIndex(), columns);
+    }
+
+    @Override
+    public DataFrame compactFloat(float forNull) {
+        int w = source.width();
+        Series<?>[] columns = new Series[w];
+
+        for (int i = 0; i < w; i++) {
+            Series s = source.getColumn(i);
+            columns[i] = s.compactFloat(forNull);
+        }
+
+        return new ColumnDataFrame(null, source.getColumnsIndex(), columns);
+    }
+
+    @Override
+    public <V> DataFrame compactFloat(FloatValueMapper<V> mapper) {
+        int w = source.width();
+        Series<?>[] columns = new Series[w];
+
+        for (int i = 0; i < w; i++) {
+            Series s = source.getColumn(i);
+            columns[i] = s.compactFloat(mapper);
         }
 
         return new ColumnDataFrame(null, source.getColumnsIndex(), columns);
@@ -287,11 +308,9 @@ public class DeferredColumnSet implements ColumnSet {
         int w = source.width();
         Series<?>[] columns = new Series[w];
 
-        DoubleValueMapper<?> mapper = DoubleValueMapper.of(forNull);
-
         for (int i = 0; i < w; i++) {
             Series s = source.getColumn(i);
-            columns[i] = s instanceof DoubleSeries ? s.castAsDouble() : s.mapAsDouble(mapper);
+            columns[i] = s.compactDouble(forNull);
         }
 
         return new ColumnDataFrame(null, source.getColumnsIndex(), columns);
@@ -304,7 +323,7 @@ public class DeferredColumnSet implements ColumnSet {
 
         for (int i = 0; i < w; i++) {
             Series s = source.getColumn(i);
-            columns[i] = s.mapAsDouble(mapper);
+            columns[i] = s.compactDouble(mapper);
         }
 
         return new ColumnDataFrame(null, source.getColumnsIndex(), columns);