+++ dflib/src/test/java/org/dflib/unit/DoubleSeriesAsserts.java
@@ -7,13 +7,20 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class DoubleSeriesAsserts {
 
-    private double[] data;
+    private final double[] data;
+    private double assertionDelta;
 
     public DoubleSeriesAsserts(DoubleSeries series) {
         assertNotNull(series, "Series is null");
 
         this.data = new double[series.size()];
         series.copyToDouble(data, 0, 0, series.size());
+        this.assertionDelta = 0.000000001;
+    }
+
+    public DoubleSeriesAsserts delta(double assertionDelta) {
+        this.assertionDelta = assertionDelta;
+        return this;
     }
 
     public DoubleSeriesAsserts expectData(double... expectedValues) {
@@ -24,7 +31,7 @@ public class DoubleSeriesAsserts {
 
             double a = data[i];
             double e = expectedValues[i];
-            assertEquals(e, a, 0.000000001, "Unexpected value at " + i);
+            assertEquals(e, a, assertionDelta, "Unexpected value at " + i);
         }
 
         return this;