+++ RELEASE-NOTES.md
@@ -24,6 +24,7 @@
 * #425 Series.replace(int,T)
 * #426 All "Json.load(..)" and "Json.save(..)" methods must be static
 * #427 Type-specific versions of Exp.$val(..)
+* #432 Unify map-to-primitive-Series API with ColumnSets
 * #433 Primitive ValueMappers consistency
 
 ## Release 1.2.0